my-python-project/OCR_Text_Recognition/requirements.txt
fastapi
uvicorn[standard]
pillow
paddleocr
opencv-python
phonenumbers
python-multipart
langdetect
numpy